+++ cubrid/CMakeLists.txt
@@ -63,6 +63,7 @@ set(THREAD_HEADERS
   ${THREAD_DIR}/thread_manager.hpp
   ${THREAD_DIR}/thread_waiter.hpp
   ${THREAD_DIR}/thread_worker_pool.hpp
+  ${THREAD_DIR}/thread_worker_pool_taskcap.hpp
   )
 
 set(BASE_SOURCES
